Fix errors thrown by `.pipe()`
