+++ lib/async.js
@@ -8,11 +8,20 @@ import {handleInputAsync, pipeOutputAsync} from './stdio/async.js';
 import {spawnedKill} from './exit/kill.js';
 import {cleanupOnExit} from './exit/cleanup.js';
 import {pipeToProcess} from './pipe/setup.js';
+import {PROCESS_OPTIONS} from './pipe/validate.js';
 import {makeAllStream} from './stream/all.js';
 import {getSpawnedResult} from './stream/resolve.js';
 import {mergePromise} from './promise.js';
 
 export const execa = (rawFile, rawArgs, rawOptions) => {
+	const {spawned, promise, options, stdioStreamsGroups} = runExeca(rawFile, rawArgs, rawOptions);
+	spawned.pipe = pipeToProcess.bind(undefined, {source: spawned, sourcePromise: promise, stdioStreamsGroups});
+	mergePromise(spawned, promise);
+	PROCESS_OPTIONS.set(spawned, options);
+	return spawned;
+};
+
+const runExeca = (rawFile, rawArgs, rawOptions) => {
 	const {file, args, command, escapedCommand, options} = handleAsyncArguments(rawFile, rawArgs, rawOptions);
 	const stdioStreamsGroups = handleInputAsync(options);
 
@@ -34,9 +43,7 @@ export const execa = (rawFile, rawArgs, rawOptions) => {
 	spawned.all = makeAllStream(spawned, options);
 
 	const promise = handlePromise({spawned, options, stdioStreamsGroups, originalStreams, command, escapedCommand, controller});
-	spawned.pipe = pipeToProcess.bind(undefined, {source: spawned, sourcePromise: promise, stdioStreamsGroups, options});
-	mergePromise(spawned, promise);
-	return spawned;
+	return {spawned, promise, options, stdioStreamsGroups};
 };
 
 const handleAsyncArguments = (rawFile, rawArgs, rawOptions) => {
