+++ lib/pipe/abort.js
@@ -1,13 +1,13 @@
 import {aborted} from 'node:util';
-import {createNonCommandError} from './validate.js';
+import {createNonCommandError} from './throw.js';
 
 export const unpipeOnAbort = (signal, ...args) => signal === undefined
 	? []
 	: [unpipeOnSignalAbort(signal, ...args)];
 
-const unpipeOnSignalAbort = async (signal, sourceStream, mergedStream, {stdioStreamsGroups, options}) => {
+const unpipeOnSignalAbort = async (signal, {sourceStream, mergedStream, stdioStreamsGroups, sourceOptions}) => {
 	await aborted(signal, sourceStream);
 	await mergedStream.remove(sourceStream);
 	const error = new Error('Pipe cancelled by `signal` option.');
-	throw createNonCommandError({error, stdioStreamsGroups, options});
+	throw createNonCommandError({error, stdioStreamsGroups, sourceOptions});
 };
