+++ lib/pipe/setup.js
@@ -1,31 +1,46 @@
 import {normalizePipeArguments} from './validate.js';
+import {handlePipeArgumentsError} from './throw.js';
 import {waitForBothProcesses} from './sequence.js';
 import {pipeProcessStream} from './streaming.js';
 import {unpipeOnAbort} from './abort.js';
 
 // Pipe a process' `stdout`/`stderr`/`stdio` into another process' `stdin`
-export const pipeToProcess = (sourceInfo, destination, pipeOptions) => {
-	const promise = handlePipePromise(sourceInfo, destination, pipeOptions);
+export const pipeToProcess = (sourceInfo, ...args) => {
+	const {destination, ...normalizedInfo} = normalizePipeArguments(sourceInfo, ...args);
+	const promise = handlePipePromise({...normalizedInfo, destination});
 	promise.pipe = pipeToProcess.bind(undefined, {...sourceInfo, source: destination, sourcePromise: promise});
 	return promise;
 };
 
-const handlePipePromise = async (sourceInfo, destination, {from, signal} = {}) => {
-	const {sourcePromise, sourceStream, destinationStream} = normalizePipeArguments(destination, from, sourceInfo);
+const handlePipePromise = async ({
+	sourcePromise,
+	sourceStream,
+	sourceOptions,
+	sourceError,
+	destination,
+	destinationStream,
+	destinationError,
+	signal,
+	stdioStreamsGroups,
+}) => {
+	handlePipeArgumentsError({
+		sourcePromise,
+		sourceStream,
+		sourceError,
+		destination,
+		destinationStream,
+		destinationError,
+		stdioStreamsGroups,
+		sourceOptions,
+	});
 	const maxListenersController = new AbortController();
 	try {
 		const mergedStream = pipeProcessStream(sourceStream, destinationStream, maxListenersController);
 		return await Promise.race([
 			waitForBothProcesses(sourcePromise, destination),
-			...unpipeOnAbort(signal, sourceStream, mergedStream, sourceInfo),
+			...unpipeOnAbort(signal, {sourceStream, mergedStream, sourceOptions, stdioStreamsGroups}),
 		]);
 	} finally {
 		maxListenersController.abort();
 	}
-};
-
-export const dummyPipeToProcess = (source, destination) => {
-	const promise = Promise.all([source, destination]);
-	promise.pipe = dummyPipeToProcess.bind(undefined, promise);
-	return promise;
 };
