+++ lib/pipe/validate.js
@@ -1,38 +1,58 @@
-import {STANDARD_STREAMS_ALIASES, isExecaChildProcess} from '../utils.js';
-import {makeEarlyError} from '../return/error.js';
-import {abortSourceStream, endDestinationStream} from './streaming.js';
+import {STANDARD_STREAMS_ALIASES} from '../utils.js';
 
-export const normalizePipeArguments = (destination, from, {source, sourcePromise, stdioStreamsGroups, options}) => {
-	const {destinationStream, destinationError} = getDestinationStream(destination);
-	const {sourceStream, sourceError} = getSourceStream(source, stdioStreamsGroups, from, options);
-	handlePipeArgumentsError({sourceStream, sourceError, destinationStream, destinationError, stdioStreamsGroups, options});
-	return {source, sourcePromise, sourceStream, destinationStream};
+export const normalizePipeArguments = ({source, sourcePromise, stdioStreamsGroups}, ...args) => {
+	const sourceOptions = PROCESS_OPTIONS.get(source);
+	const {
+		destination,
+		destinationStream,
+		destinationError,
+		pipeOptions: {from, signal} = {},
+	} = getDestinationStream(...args);
+	const {sourceStream, sourceError} = getSourceStream(source, stdioStreamsGroups, from, sourceOptions);
+	return {
+		sourcePromise,
+		sourceStream,
+		sourceOptions,
+		sourceError,
+		destination,
+		destinationStream,
+		destinationError,
+		signal,
+		stdioStreamsGroups,
+	};
 };
 
-const getDestinationStream = destination => {
+const getDestinationStream = (...args) => {
 	try {
-		if (!isExecaChildProcess(destination)) {
-			throw new TypeError(`The first argument must be an Execa child process: ${destination}`);
-		}
-
+		const {destination, pipeOptions} = getDestination(...args);
 		const destinationStream = destination.stdin;
 		if (destinationStream === null) {
 			throw new TypeError('The destination child process\'s stdin must be available. Please set its "stdin" option to "pipe".');
 		}
 
-		return {destinationStream};
+		return {destination, destinationStream, pipeOptions};
 	} catch (error) {
 		return {destinationError: error};
 	}
 };
 
-const getSourceStream = (source, stdioStreamsGroups, from, options) => {
+const getDestination = (firstArgument, ...args) => {
+	if (!PROCESS_OPTIONS.has(firstArgument)) {
+		throw new TypeError(`The first argument must be an Execa child process: ${firstArgument}`);
+	}
+
+	return {destination: firstArgument, pipeOptions: args[0]};
+};
+
+export const PROCESS_OPTIONS = new WeakMap();
+
+const getSourceStream = (source, stdioStreamsGroups, from, sourceOptions) => {
 	try {
 		const streamIndex = getStreamIndex(stdioStreamsGroups, from);
 		const sourceStream = streamIndex === 'all' ? source.all : source.stdio[streamIndex];
 
 		if (sourceStream === null || sourceStream === undefined) {
-			throw new TypeError(getInvalidStdioOptionMessage(streamIndex, from, options));
+			throw new TypeError(getInvalidStdioOptionMessage(streamIndex, from, sourceOptions));
 		}
 
 		return {sourceStream};
@@ -73,12 +93,12 @@ Please set the "stdio" option to ensure that file descriptor exists.`);
 	return streamIndex;
 };
 
-const getInvalidStdioOptionMessage = (streamIndex, from, options) => {
-	if (streamIndex === 'all' && !options.all) {
+const getInvalidStdioOptionMessage = (streamIndex, from, sourceOptions) => {
+	if (streamIndex === 'all' && !sourceOptions.all) {
 		return 'The "all" option must be true to use `childProcess.pipe(destinationProcess, {from: "all"})`.';
 	}
 
-	const {optionName, optionValue} = getInvalidStdioOption(streamIndex, options);
+	const {optionName, optionValue} = getInvalidStdioOption(streamIndex, sourceOptions);
 	const pipeOptions = from === undefined ? '' : `, {from: ${serializeOptionValue(from)}}`;
 	return `The \`${optionName}: ${serializeOptionValue(optionValue)}\` option is incompatible with using \`childProcess.pipe(destinationProcess${pipeOptions})\`.
 Please set this option with "pipe" instead.`;
@@ -105,36 +125,3 @@ const serializeOptionValue = optionValue => {
 
 	return typeof optionValue === 'number' ? `${optionValue}` : 'Stream';
 };
-
-const handlePipeArgumentsError = ({sourceStream, sourceError, destinationStream, destinationError, stdioStreamsGroups, options}) => {
-	const error = getPipeArgumentsError({sourceStream, sourceError, destinationStream, destinationError});
-	if (error !== undefined) {
-		throw createNonCommandError({error, stdioStreamsGroups, options});
-	}
-};
-
-const getPipeArgumentsError = ({sourceStream, sourceError, destinationStream, destinationError}) => {
-	if (sourceError !== undefined && destinationError !== undefined) {
-		return destinationError;
-	}
-
-	if (destinationError !== undefined) {
-		abortSourceStream(sourceStream);
-		return destinationError;
-	}
-
-	if (sourceError !== undefined) {
-		endDestinationStream(destinationStream);
-		return sourceError;
-	}
-};
-
-export const createNonCommandError = ({error, stdioStreamsGroups, options}) => makeEarlyError({
-	error,
-	command: PIPE_COMMAND_MESSAGE,
-	escapedCommand: PIPE_COMMAND_MESSAGE,
-	stdioStreamsGroups,
-	options,
-});
-
-const PIPE_COMMAND_MESSAGE = 'source.pipe(destination)';
