+++ lib/script.js
@@ -1,7 +1,8 @@
 import isPlainObject from 'is-plain-obj';
-import {isBinary, binaryToString, isChildProcess, isExecaChildProcess} from './utils.js';
+import {isBinary, binaryToString, isChildProcess} from './utils.js';
 import {execa} from './async.js';
 import {execaSync} from './sync.js';
+import {PROCESS_OPTIONS} from './pipe/validate.js';
 
 const create$ = options => {
 	function $(templatesOrOptions, ...expressions) {
@@ -46,7 +47,7 @@ const scriptWithPipe = returnValue => {
 };
 
 const scriptPipe = (pipeMethod, options, firstArgument, ...args) => {
-	if (isExecaChildProcess(firstArgument)) {
+	if (PROCESS_OPTIONS.has(firstArgument)) {
 		if (Object.keys(options).length > 0) {
 			throw new TypeError('Please use .pipe(options)`command` or .pipe($(options)`command`) instead of .pipe(options)($`command`).');
 		}
