+++ lib/utils.js
@@ -27,8 +27,4 @@ export const incrementMaxListeners = (eventEmitter, maxListenersIncrement, signa
 	});
 };
 
-export const isExecaChildProcess = value => isChildProcess(value)
-	&& typeof value.then === 'function'
-	&& typeof value.pipe === 'function';
-
 export const isChildProcess = value => value instanceof ChildProcess;
